pyproject.toml
[tool.poetry]
name = "openai-assistant"
version = "0.1.0"
description = ""
authors = ["JPDas <jprakash.das@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
openai = "^1.55.1"
streamlit = "^1.40.2"
pdfplumber = "^0.11.4"
docx = "^0.2.4"
langchain = "^0.3.9"
chromadb = "^0.5.20"
openpyxl = "^3.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
